backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow==2.20.0
numpy==2.2.6
opencv-python==4.12.0.88
python-multipart==0.0.6
pillow==11.3.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.10
alembic==1.14.0
